verify user before delete post

+++ app/controllers/Posts.php
@@ -128,6 +128,14 @@
 
         public function delete($id){
             if($_SERVER['REQUEST_METHOD']== 'POST'){
+
+                //get existing post from model
+                $post = $this->postModel->getPostById($id);
+                //check for owner
+                if($post->user_id != $_SESSION['user_id']){
+                    redirect('posts');
+                }
+                
                 if($this->postModel->deletePost($id)){
                     flash('post_message', 'Post Deleted');
                     redirect('posts');
